* commit '9c2110a41019e6c8daab8e3bfe6d0f9d9c2fafb3':
  KERNEL-595 FineIO 2.0、3.0接口统一
  KERNEL-595 FineIO 2.0、3.0接口统一
  KERNEL-595 FineIO 2.0、3.0接口统一
  KERNEL-595 FineIO 2.0、3.0接口统一
  KERNEL-595 FineIO 2.0、3.0接口统一

+++ src/main/java/com/fineio/v3/connector/BaseConnector.java
@@ -1,27 +1,29 @@
 package com.fineio.v3.connector;
 
+import com.fineio.storage.Connector;
+
 /**
  * @author yee
  */
 public abstract class BaseConnector implements Connector {
 
-    private final int blockOffset;
+    private final byte blockOffset;
 
     /**
      * Default constructor
      *
      * @param blockOffset
      */
-    public BaseConnector(int blockOffset) {
+    public BaseConnector(byte blockOffset) {
         this.blockOffset = blockOffset;
     }
 
     public BaseConnector() {
-        this(22);
+        this((byte) 22);
     }
 
     @Override
-    public int getBlockOffset() {
+    public byte getBlockOffset() {
         return blockOffset;
     }
 }
